+++ Dashboard/src/main/java/com/savit/dashboard/presentation/viewstate/DashboardViewEvent.kt
@@ -4,7 +4,7 @@ import com.savit.core.base.viewstate.ViewEvent
 
 sealed class DashboardViewEvent : ViewEvent {
 
-    class OpenRecord(val recordId: Long) : DashboardViewEvent()
+    class OpenRecord(val recordId: Int) : DashboardViewEvent()
 
     object AddRecord : DashboardViewEvent()
 }
